Glob: Rename to Globbing
I think it makes sense to have all function chunks named ending with
'ing' and all objects objecty.
