+++ coalib/parsing/Globbing.py
